refactor: tce broadcast config
