+++ crates/topos-config/Cargo.toml
@@ -7,11 +7,7 @@ edition = "2021"
 
 [dependencies]
 topos-p2p = { path = "../topos-p2p" }
-topos-tce-transport = { path = "../topos-tce-transport" }
-topos-sequencer = { path = "../topos-sequencer" }
-topos-core = { workspace = true, features = ["api"] }
-topos-certificate-spammer = { path = "../topos-certificate-spammer" }
-topos-tce-broadcast = { path = "../topos-tce-broadcast", optional = true }
+topos-core = { path = "../topos-core" }
 topos-wallet = { path = "../topos-wallet" }
 
 async-stream.workspace = true
@@ -48,7 +44,6 @@ openssl = { version = "0.10.61", features = ["vendored"] }
 
 [dev-dependencies]
 topos-tce-broadcast = { path = "../topos-tce-broadcast" }
-topos-tce-transport = { path = "../topos-tce-transport" }
 topos-tce-synchronizer = { path = "../topos-tce-synchronizer" }
 topos-tce-gatekeeper = { path = "../topos-tce-gatekeeper" }
 topos-tce-api = { path = "../topos-tce-api" }
