+++ crates/topos-tce-broadcast/src/double_echo/broadcast_state.rs
@@ -1,7 +1,7 @@
+use crate::event::ProtocolEvents;
 use crate::sampler::SubscriptionsView;
 use std::sync::Arc;
 use std::{collections::HashSet, time};
-use tce_transport::ProtocolEvents;
 use tokio::sync::mpsc;
 use topos_core::{
     types::{
