+++ crates/topos-tce-broadcast/src/event.rs
@@ -1,38 +1,8 @@
-//! implementation of Topos Network Transport
-//!
-use clap::Parser;
-use serde::{Deserialize, Serialize};
 use topos_core::{
     types::ValidatorId,
     uci::{Certificate, CertificateId},
 };
 use topos_crypto::messages::Signature;
-
-#[derive(Parser, Clone, Debug, Default, Deserialize, Serialize)]
-#[command(name = "Parameters of the reliable broadcast")]
-pub struct ReliableBroadcastParams {
-    /// Echo threshold
-    #[arg(long, env = "TCE_ECHO_THRESHOLD", default_value_t = 1)]
-    pub echo_threshold: usize,
-    /// Ready threshold
-    #[arg(long, env = "TCE_READY_THRESHOLD", default_value_t = 1)]
-    pub ready_threshold: usize,
-    /// Delivery threshold
-    #[arg(long, env = "TCE_DELIVERY_THRESHOLD", default_value_t = 1)]
-    pub delivery_threshold: usize,
-}
-
-impl ReliableBroadcastParams {
-    pub fn new(n: usize) -> Self {
-        let f: usize = n / 3;
-
-        Self {
-            echo_threshold: 1 + (n + f) / 2,
-            ready_threshold: 1 + f,
-            delivery_threshold: 2 * f + 1,
-        }
-    }
-}
 
 /// Protocol events
 #[derive(Clone, Debug)]
@@ -66,7 +36,4 @@ pub enum ProtocolEvents {
     },
     /// For simulation purpose, for now only caused by ill-formed sampling
     Die,
-
-    /// Stable Sample
-    StableSample,
 }
