+++ crates/topos-tce-broadcast/src/tests/mod.rs
@@ -4,8 +4,8 @@ use rstest::*;
 use std::collections::HashSet;
 use std::str::FromStr;
 use std::time::Duration;
-use tce_transport::ReliableBroadcastParams;
 use tokio::sync::mpsc::Receiver;
+use topos_config::tce::broadcast::ReliableBroadcastParams;
 use topos_crypto::messages::MessageSigner;
 use topos_test_sdk::constants::*;
 use topos_test_sdk::storage::create_validator_store;
