--- crates/topos-tce-transport/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "topos-tce-transport"
-version = "0.1.0"
-edition = "2021"
-
-[lints]
-workspace = true
-
-[lib]
-path = 'src/lib.rs'
-
-[dependencies]
-clap.workspace = true
-serde = { workspace = true, features = ["rc"] }
-tracing.workspace = true
-topos-core = { workspace = true, features = ["api", "uci"] }
-topos-crypto = { path = "../topos-crypto" }
