+++ crates/topos-tce/Cargo.toml
@@ -31,7 +31,6 @@ tonic.workspace = true
 bytes.workspace = true
 prost.workspace = true
 
-tce_transport = { package = "topos-tce-transport", path = "../topos-tce-transport" }
 topos-config = { path = "../topos-config" }
 topos-p2p = { path = "../topos-p2p" }
 topos-metrics = { path = "../topos-metrics" }
