+++ crates/topos-tce/src/app_context.rs
@@ -6,7 +6,6 @@ use futures::{Stream, StreamExt};
 use prometheus::HistogramTimer;
 use std::collections::HashMap;
 use std::sync::Arc;
-use tce_transport::ProtocolEvents;
 use tokio::sync::mpsc;
 use tokio_util::sync::CancellationToken;
 use topos_core::uci::CertificateId;
@@ -14,6 +13,7 @@ use topos_metrics::CERTIFICATE_DELIVERED_TOTAL;
 use topos_p2p::{Event as NetEvent, NetworkClient};
 use topos_tce_api::RuntimeClient as ApiClient;
 use topos_tce_api::RuntimeEvent as ApiEvent;
+use topos_tce_broadcast::event::ProtocolEvents;
 use topos_tce_broadcast::ReliableBroadcastClient;
 use topos_tce_gatekeeper::GatekeeperClient;
 use topos_tce_storage::store::ReadStore;
