+++ crates/topos-test-sdk/src/tce/protocol.rs
@@ -2,12 +2,13 @@ use futures::Stream;
 use std::collections::HashSet;
 use std::sync::Arc;
 use tokio::sync::broadcast;
+use topos_config::tce::broadcast::ReliableBroadcastParams;
 use topos_core::types::ValidatorId;
 use topos_crypto::messages::MessageSigner;
+use topos_tce_broadcast::event::ProtocolEvents;
 use topos_tce_broadcast::{ReliableBroadcastClient, ReliableBroadcastConfig};
 use topos_tce_storage::types::CertificateDeliveredWithPositions;
 use topos_tce_storage::validator::ValidatorStore;
-use topos_tce_transport::{ProtocolEvents, ReliableBroadcastParams};
 
 pub async fn create_reliable_broadcast_client(
     validator_id: ValidatorId,
